Started adding combat encounters
I started the process of adding combat encounters to the game. When the player is in a dungeon, it randomly rolls every second and if the random number is below the number of enemies in the dungeon, it will switch to the combat state. I also created an Enterable class and factored out some of the functions from the Building class to the Enterable class so that it does not need to be written in the Dungeon class as well.